New Development: No             JIRA Issue:         Ready for Test: Yes      Interface Changes: No What Interface Changed:          Test Programs:   Put in Release Notes: No              Module(s): asdm2MS            Description: fix type ambiguity in min(T, V)
git-svn-id: https://svn.cv.nrao.edu/svn/casa/trunk@37793 ac9c0da5-131f-0410-9009-a9c195f183e6

+++ code/alma/apps/asdm2MS/asdm2MS.cc
@@ -7313,7 +7313,7 @@ int main(int argc, char *argv[]) {
 	  
 	  // For each slice of the BDF with a size approx equal to the required size
 	  for (unsigned int j = 0; j < actualSizeInMemory.size(); j++) {
-	    numberOfIntegrations = min(actualSizeInMemory[j] / (bdfSize / N), N); // The min to prevent a possible excess when there are very few bytes in the BDF.
+	    numberOfIntegrations = min((uint64_t)actualSizeInMemory[j] / (bdfSize / N), (uint64_t)N); // The min to prevent a possible excess when there are very few bytes in the BDF.
 	    if (numberOfIntegrations) {
 	      infostream.str("");
 	      infostream << "ASDM Main row #" << mainRowIndex[i] << " - " << numberOfReadIntegrations  << " integrations done so far - the next " << numberOfIntegrations << " integrations produced " ;
